requirements_query_subir.txt
apache-beam==2.52.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
db-dtypes==1.2.0
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
fastavro==1.9.1
fasteners==0.19
google-api-core==2.15.0
google-apitools==0.5.31
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.14.1
google-cloud-bigquery-storage==2.24.0
google-cloud-bigtable==2.22.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-dlp==3.14.0
google-cloud-language==2.12.0
google-cloud-pubsub==2.19.0
google-cloud-pubsublite==1.9.0
google-cloud-recommendations-ai==0.10.6
google-cloud-resource-manager==1.11.0
google-cloud-spanner==3.40.1
google-cloud-storage==2.14.0
google-cloud-videointelligence==2.12.0
google-cloud-vision==3.5.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.6
Js2Py==0.74
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
numpy==1.24.4
oauth2client==4.1.3
objsize==0.6.1
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.4
proto-plus==1.23.0
protobuf==4.25.1
pyarrow==11.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydot==1.4.2
pyjsparser==2.7.1
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.2
rsa==4.9
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
zstandard==0.22.0
